+++ backend/app.js
@@ -1,53 +1,14 @@
-const app = require("./app");
-const http = require("http");
+const express = require('express');
 
-const normalizePort = val => {
-  var port = parseInt(val, 10);
+const app = express();
 
-  if (isNaN(port)) {
-    // named pipe
-    return val;
-  }
+app.use((req, res, next) => {
+    console.log("first middle ware");
+    next();
+});
 
-  if (port >= 0) {
-    // port number
-    return port;
-  }
+app.use((req, res, next) => {
+    res.send('hello from express!');
+});
 
-  return false;
-};
-
-const onError = error => {
-  if (error.syscall !== "listen") {
-    throw error;
-  }
-  const bind = typeof port === "string" ? "pipe " + port : "port " + port;
-  switch (error.code) {
-    case "EACCES":
-      console.error(bind + " requires elevated privileges");
-      process.exit(1);
-      break;
-    case "EADDRINUSE":
-      console.error(bind + " is already in use");
-      process.exit(1);
-      break;
-    default:
-      throw error;
-  }
-};
-
-const onListening = () => {
-  const addr = server.address();
-  const bind = typeof port === "string" ? "pipe " + port : "port " + port;
-  debug("Listening on " + bind);
-};
-
-
-
-const port = normalizePort(process.env.PORT || "3000");
-app.set("port", port);
-
-const server = http.createServer(app);
-server.on("error", onError);
-server.on("listening", onListening);
-server.listen(port);
+module.exports = app;
